Migrate Modal to new storybook
