+++ src/modal/Modal.story.mdx
@@ -1,0 +1,62 @@
+import { Meta, Preview, Props, Story } from '@storybook/addon-docs/blocks'
+
+import { Modal, ModalSize } from './index'
+import { ModalOpener } from './storyUtils/ModalOpener'
+
+<Meta title="Widgets|Modal" />
+
+
+# **Modal**
+
+### Default
+
+<Preview>
+  <Story name="default">
+    <ModalOpener
+      onClose={() => {}}
+      closeOnEsc
+      closeOnOutsideClick
+      displayDimmer
+      isOpen={false}
+      closeButtonTitle="Close modal"
+      aria-labelledby="label1"
+      aria-describedby="description1"
+      size={ModalSize.MEDIUM}
+      isLoading={false}
+    />
+  </Story>
+</Preview>
+
+### FullScreen
+
+<Preview>
+  <Story name="fullscreen">
+    <ModalOpener
+      onClose={() => {}}
+      closeOnEsc
+      closeOnOutsideClick={false}
+      displayDimmer={false}
+      displayCloseButton
+      isOpen={false}
+      aria-labelledby="label2"
+      aria-describedby="description2"
+      size={ModalSize.FULLSCREEN}
+      isLoading={false}
+    />
+  </Story>
+</Preview>
+
+## Specifications
+
+WAI ARIA: https://www.w3.org/TR/wai-aria-practices-1.1/#dialog
+
+
+## Usage
+
+```jsx
+import { Modal } from '@blablacar/ui-library/build/modal'
+
+<Modal onClose={() => null} isOpen isLoading={false} />
+```
+
+<Props of={Modal} />
