--- src/modal/specifications/modal.md
@@ -1,13 +1,0 @@
-# Modal
-
-```js
-import { Modal } from '@blablacar/ui-library'
-```
-
-## Playground
-
-<!-- STORY -->
-
-## Specifications
-
-WAI ARIA: https://www.w3.org/TR/wai-aria-practices-1.1/#dialog
